refactor: Header, Dropdown 컴포넌트 [#2]
