+++ src/App.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { BrowserRouter, Routes, Route} from 'react-router-dom';
 import Header from './Header';
 import Footer from './Footer';
