+++ src/component/Product.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import './product.css';
 
 export default function Product() {
