+++ src/page/Bookmark.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 
 
 export default function Bookmark() {
